requirements.txt
flask>=2.3.0
pyyaml>=6.0
click>=8.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.16.0
google-generativeai>=0.8.0
jinja2>=3.1.0
rich>=13.0.0
tabulate>=0.9.0
python-dateutil>=2.8.0
colorama>=0.4.6
prompt-toolkit>=3.0.0
faker>=20.0.0
gunicorn>=21.0.0
